requirements.txt
Django==5.2.6
Pillow==10.1.0
python-dotenv==1.0.0
